Fix race condition, accessibility, stacking, and button text in confirm dialogs

+++ components/annotorious-annotator/line-parser.js
@@ -20,6 +20,7 @@ import { CleanupRegistry } from '../../utilities/CleanupRegistry.js'
 import { onProjectReady } from '../../utilities/projectReady.js'
 import vault from '../../js/vault.js'
 import '../page-selector/index.js'
+import { confirmAction } from '../../utilities/confirmAction.js'
 
 class AnnotoriousAnnotator extends HTMLElement {
   #osd
@@ -715,16 +716,10 @@ class AnnotoriousAnnotator extends HTMLElement {
         srcDown: "../interfaces/annotator/images/transcribe.png",
         onClick: (e) => {
             if (this.#resolvedAnnotationPage?.$isDirty) {
-              const onPositive = () => {
-                TPEN.eventDispatcher.off('tpen-confirm-negative', onNegative)
-                location.href = `/transcribe?projectID=${TPEN.activeProject._id}&pageID=${this.#annotationPageID}`
-              }
-              const onNegative = () => {
-                TPEN.eventDispatcher.off('tpen-confirm-positive', onPositive)
-              }
-              TPEN.eventDispatcher.one('tpen-confirm-positive', onPositive)
-              TPEN.eventDispatcher.one('tpen-confirm-negative', onNegative)
-              TPEN.eventDispatcher.dispatch('tpen-confirm', { message: "Stop identifying lines and go transcribe?  Unsaved changes will be lost." })
+              confirmAction(
+                "Stop identifying lines and go transcribe?  Unsaved changes will be lost.",
+                () => { location.href = `/transcribe?projectID=${TPEN.activeProject._id}&pageID=${this.#annotationPageID}` }
+              )
             }
             else {
               location.href = `/transcribe?projectID=${TPEN.activeProject._id}&pageID=${this.#annotationPageID}`  
@@ -816,18 +811,14 @@ class AnnotoriousAnnotator extends HTMLElement {
           _this.#pendingTimeouts.delete(timeoutId)
           // Timeout required in order to allow the click-and-focus native functionality to complete.
           // Also stops the goofy UX for naturally slow clickers.
-          const onPositive = () => {
-            TPEN.eventDispatcher.off('tpen-confirm-negative', onNegative)
-            _this.#annotoriousInstance.removeAnnotation(originalAnnotation)
-            _this.#resolvedAnnotationPage.$isDirty = true
-          }
-          const onNegative = () => {
-            TPEN.eventDispatcher.off('tpen-confirm-positive', onPositive)
-            _this.#annotoriousInstance.cancelSelected()
-          }
-          TPEN.eventDispatcher.one('tpen-confirm-positive', onPositive)
-          TPEN.eventDispatcher.one('tpen-confirm-negative', onNegative)
-          TPEN.eventDispatcher.dispatch('tpen-confirm', { message: "Are you sure you want to remove this?" })
+          confirmAction(
+            "Are you sure you want to remove this?",
+            () => {
+              _this.#annotoriousInstance.removeAnnotation(originalAnnotation)
+              _this.#resolvedAnnotationPage.$isDirty = true
+            },
+            () => _this.#annotoriousInstance.cancelSelected()
+          )
         }, 500)
         _this.#pendingTimeouts.add(timeoutId)
       }
@@ -1157,33 +1148,31 @@ class AnnotoriousAnnotator extends HTMLElement {
    * https://annotorious.dev/api-reference/openseadragon-annotator/#clearannotations
    */
   deleteAllAnnotations() {
-    const onPositive = () => {
-      TPEN.eventDispatcher.off('tpen-confirm-negative', onNegative)
-      const deleteAllBtn = this.shadowRoot.getElementById("deleteAllBtn")
-      deleteAllBtn.setAttribute("disabled", "true")
-      deleteAllBtn.textContent = "deleting.  please wait..."
-      this.#annotoriousInstance.clearAnnotations()
-      this.#resolvedAnnotationPage.$isDirty = true
-      this.saveAnnotations()
-        .then(() => this.clearColumnsServerSide())
-        .catch(err => {
-          console.error("Could not delete all annotations.", err)
-          TPEN.eventDispatcher.dispatch("tpen-toast", {
-            message: "Could not delete all annotations.",
-            status: "error"
+    confirmAction(
+      'This will remove all Annotations and will take effect immediately. This action cannot be undone.',
+      () => {
+        const deleteAllBtn = this.shadowRoot.getElementById("deleteAllBtn")
+        deleteAllBtn.setAttribute("disabled", "true")
+        deleteAllBtn.textContent = "deleting.  please wait..."
+        this.#annotoriousInstance.clearAnnotations()
+        this.#resolvedAnnotationPage.$isDirty = true
+        this.saveAnnotations()
+          .then(() => this.clearColumnsServerSide())
+          .catch(err => {
+            console.error("Could not delete all annotations.", err)
+            TPEN.eventDispatcher.dispatch("tpen-toast", {
+              message: "Could not delete all annotations.",
+              status: "error"
+            })
           })
-        })
-        .finally(() => {
-          deleteAllBtn.removeAttribute("disabled")
-          deleteAllBtn.textContent = "Delete All Annotations"
-        })
-    }
-    const onNegative = () => {
-      TPEN.eventDispatcher.off('tpen-confirm-positive', onPositive)
-    }
-    TPEN.eventDispatcher.one('tpen-confirm-positive', onPositive)
-    TPEN.eventDispatcher.one('tpen-confirm-negative', onNegative)
-    TPEN.eventDispatcher.dispatch('tpen-confirm', { message: 'This will remove all Annotations and will take effect immediately. This action cannot be undone.' })
+          .finally(() => {
+            deleteAllBtn.removeAttribute("disabled")
+            deleteAllBtn.textContent = "Delete All Annotations"
+          })
+      },
+      null,
+      { positiveButtonText: "Delete All", negativeButtonText: "Cancel" }
+    )
   }
 
   /**
